test(io): move CORS related tests to test_cors.py
Moving them to its own module makes sense. The goal is to group related
tests.
